+++ test/runtests.jl
@@ -49,6 +49,7 @@ include("large_scale_condensation.jl")
 include("optical_depth.jl")
 include("longwave_radiation.jl")
 # include("shortwave_radiation.jl")
+include("surface_fluxes.jl")
 include("random_process.jl")
 include("stochastic_physics.jl")